removing old datasource interface and classes
