+++ src/main/java/com/autotune/experimentManager/handler/MetricCollectionHandler.java
@@ -18,9 +18,7 @@ package com.autotune.experimentManager.handler;
 import com.autotune.analyzer.utils.AnalyzerConstants;
 import com.autotune.common.data.metrics.Metric;
 import com.autotune.common.data.metrics.MetricResults;
-import com.autotune.common.datasource.DataSourceOperator;
 import com.autotune.common.datasource.DataSourceOperatorImpl;
-import com.autotune.common.datasource.KruizeDataSourceOperator;
 import com.autotune.common.k8sObjects.KubernetesContexts;
 import com.autotune.common.parallelengine.executor.KruizeExecutor;
 import com.autotune.common.parallelengine.worker.CallableFactory;
